动画：设置速率

+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/prefs/screens/LauncherPreferenceExclusiveFragment.java
@@ -15,20 +15,20 @@ public class LauncherPreferenceExclusiveFragment extends LauncherPreferenceFragm
         addPreferencesFromResource(R.xml.pref_exclusive);
 
         // 获取值
-        int rate = PREF_ANIMATION_RATE;
+        long rate = PREF_ANIMATION_RATE;
 
         CustomSeekBarPreference seek = requirePreference("animationRate", CustomSeekBarPreference.class);
         seek.setMin(0);
         seek.setMax(1000);
-        seek.setValue(rate);
+        seek.setValue((int) rate);
         seek.setSuffix(" ms");
 
         seek.setOnPreferenceChangeListener((preference, newValue) -> {
             Animation cutInto = AnimationUtils.loadAnimation(requireContext(), R.anim.cut_into);
             Animation cutOut = AnimationUtils.loadAnimation(requireContext(), R.anim.cut_out);
 
-            cutInto.setDuration((Integer) newValue);
-            cutOut.setDuration((Integer) newValue);
+            cutInto.setDuration((Long) newValue);
+            cutOut.setDuration((Long) newValue);
             return true;
         });
     }
